bug fix for two things (Tor Lillqvist)
1) This is probably not important: Add a function
must_use_null_useddefaultchar() that checks for those codepages for
which the docs for WideCharToMultiByte() say one has to use a NULL
lpDefaultChar pointer. Don't know if this is actually needed, but
better to be safe than sorry.

2) This is essential: In kernel_wctomb(), the code should first check
if bufsize is zero, and return the E2BIG error in that case.